Partial inlining
Recursively inlines partials by making calls to the text loader plugin.
It inlines only the partials with the designated path delimiter in the
partial name (by default, `/`).

This allows the use of partials like `{{> path/to/template}}` without
having to manually do it. Configuration can be done through the
RequireJS module configuration instead of using the hgn key on the
RequireJS global configuration object.

    require.config({
      'path/to/hgn': {
        pathPrefix: 'templates/'
      }
    });